Moved session object to Main Container
