+++ src/App.js
@@ -1,10 +1,8 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
+import MainContainer from './components/MainContainer';
 import generateStore from './redux/store';
-import Sidebar from './components/Sidebar';
-import PageContent from './components/PageContent';
-import aStyle from './styles/index.module.css';
 
 const App = () => {
   const store = generateStore();
@@ -12,12 +10,7 @@ const App = () => {
   return (
     <Provider store={store}>
       <BrowserRouter>
-        <>
-          <div className={aStyle.dFlexWrapper}>
-            <Sidebar />
-            <PageContent />
-          </div>
-        </>
+        <MainContainer />
       </BrowserRouter>
     </Provider>
   );
